rewritten qofspider teardown/termination feature
thread functions are now wrapped into exception_wrapper(), which reports
eventual exceptions back to the mplane component (it has to check
EcnSpider2.exception is not None). After that, a teardown sequence is
initiated by calling terminate(), which terminates qof, closes
listeners, joins all threads and empties the various queues. Emptying
the queues is necessary, so stop() does not block infinitely due to
calls to Queue.join().